`ImageCommentsViewController` displays remote image comments when customer has connectivity
