+++ EssentialApp/EssentialAppTests/ImageCommentsUIIntegrationTests.swift
@@ -196,61 +196,3 @@ final class ImageCommentsUIIntegrationTests: XCTestCase {
 	}
 
 }
-
-private extension ImageCommentsViewController {
-	var isShowingLoadingIndicator: Bool {
-		refreshControl?.isRefreshing == true
-	}
-
-	private var imageCommentsSection: Int {
-		return 0
-	}
-
-	func simulateUserInitiatedImageCommentsReload() {
-		refreshControl?.simulatePullToRefresh()
-	}
-
-	func numberOfRenderedImageCommentViews() -> Int {
-		tableView.numberOfRows(inSection: imageCommentsSection)
-	}
-
-	var errorMessage: String? {
-		errorView?.message
-	}
-
-	@discardableResult
-	func simulateImageCommentViewVisible(at index: Int) -> ImageCommentCell? {
-		imageCommentView(at: index) as? ImageCommentCell
-	}
-
-	@discardableResult
-	func simulateImageCommentNotVisible(at row: Int) -> ImageCommentCell? {
-		let view = simulateImageCommentViewVisible(at: row)
-
-		let delegate = tableView.delegate
-		let index = IndexPath(row: row, section: imageCommentsSection)
-		delegate?.tableView?(tableView, didEndDisplaying: view!, forRowAt: index)
-
-		return view
-	}
-
-	func imageCommentView(at row: Int) -> UITableViewCell? {
-		let ds = tableView.dataSource
-		let index = IndexPath(row: row, section: imageCommentsSection)
-		return ds?.tableView(tableView, cellForRowAt: index)
-	}
-}
-
-private extension ImageCommentCell {
-	var messageText: String? {
-		messageLabel.text
-	}
-
-	var authorNameText: String? {
-		authorNameLabel.text
-	}
-
-	var createdAtText: String? {
-		createdAtLabel.text
-	}
-}
